Move components from util to own files (react/no-multi-comp)
